+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/sink/UniversalKafkaSinkFactory.scala
@@ -270,7 +270,6 @@ class UniversalKafkaSinkFactory(
         validationAgainstSchemaErrors,
         Some(state)
       )
-
     case TransformationStep(
           (`topicParamName`, DefinedEagerParameter(_: String, _)) ::
           (`contentTypeParamName`, DefinedEagerParameter(_: String, _)) ::
